add dockerfile and major improvements

+++ telegram/telegram_test.py
@@ -12,7 +12,7 @@ from aiogram.contrib.fsm_storage.memory import MemoryStorage
 from aiogram.dispatcher import FSMContext
 from aiogram.dispatcher.filters.state import State, StatesGroup
 
-from function import create_category, get_button_labels, send_category_keyboard, delete_category
+from telegram.function import create_category, get_button_labels, send_category_keyboard, delete_category
 
 API_TOKEN = config('API_TOKEN')
 storage = MemoryStorage()
@@ -57,20 +57,21 @@ async def get_all_tasks(chat_id, message_id=None, user_id=None, chat_type=None, 
     async with aiohttp.ClientSession() as session:
         async with session.get(url) as resp:
             tasks = await resp.json()
+    if len(tasks) > 0:
+        buttons = [
+            [InlineKeyboardButton(text=task.get('title'), callback_data=f"delete_task:{task.get('id')}:{tag}")]
+            for task in tasks
+        ]
 
-    buttons = [
-        [InlineKeyboardButton(text=task.get('title'), callback_data=f"delete_task:{task.get('id')}:{tag}")]
-        for task in tasks
-    ]
-
-    keyboard = InlineKeyboardMarkup(inline_keyboard=buttons)
-    if chat_type == 'private':
-        await bot.send_message(chat_id, 'Ваши задачи', reply_markup=keyboard)
-    elif message_id:
-        await bot.edit_message_reply_markup(chat_id=chat_id, message_id=message_id, reply_markup=keyboard)
+        keyboard = InlineKeyboardMarkup(inline_keyboard=buttons)
+        if chat_type == 'private':
+            await bot.send_message(chat_id, 'Ваши задачи', reply_markup=keyboard)
+        elif message_id:
+            await bot.edit_message_reply_markup(chat_id=chat_id, message_id=message_id, reply_markup=keyboard)
+        else:
+            await bot.send_message(chat_id, 'Ваши задачи', reply_markup=keyboard)
     else:
-        await bot.send_message(chat_id, 'Ваши задачи', reply_markup=keyboard)
-
+        await bot.send_message(chat_id, 'У вас нет задач')
 
 @dp.message_handler(commands=['start'])
 async def start_handler(message: types.Message):
@@ -118,11 +119,6 @@ async def is_in_buttons(user_id: int, text: str):
     return text in button_labels
 
 
-@dp.message_handler(lambda message: asyncio.ensure_future(is_in_buttons(message.from_user.id, message.text)))
-async def handle_category_click_wrapper(message: types.Message):
-    await handle_category_click(message)
-
-
 async def handle_category_click(message: types.Message):
     tag = message.text
     await get_all_tasks(chat_id=message.chat.id, message_id=message.message_id, user_id=message.from_user.id, tag=tag,
@@ -135,15 +131,20 @@ async def create_category_handler(message: types.Message):
     await CreateCategory.name.set()
 
 
+# todo не пускать пустые задачи
 @dp.message_handler(state=CreateCategory.name)
 async def process_create_category(message: types.Message, state: FSMContext):
-    category_name = message.text
-    error_msg = await create_category(message.from_user.id, category_name)
-    if error_msg is not None:
-        await message.answer(error_msg)
-    await state.finish()
-    keyboard = await send_category_keyboard(message.from_user.id)
-    await bot.send_message(chat_id=message.chat.id, text="Выберите категорию:", reply_markup=keyboard)
+    if message.text == 'Home':
+        await bot.send_message(message.chat.id, "Главное меню", reply_markup=main_keyboard)
+    else:
+        category_name = message.text
+        error_msg = await create_category(message.from_user.id, category_name)
+        if error_msg is not None:
+            await message.answer(error_msg)
+        else:
+            await state.finish()
+        keyboard = await send_category_keyboard(message.from_user.id)
+        await bot.send_message(chat_id=message.chat.id, text="Выберите категорию:", reply_markup=keyboard)
 
 
 @dp.message_handler(lambda message: message.text == 'Удаление категории')
@@ -153,14 +154,24 @@ async def handle_delete_category(message: types.Message, state: FSMContext) -> N
     Args:
         message: Message object from the user.
     """
-    try:
-        keyboard = ReplyKeyboardMarkup(resize_keyboard=True, row_width=4, selective=True).add(
-            *(KeyboardButton(text=label) for label in await get_button_labels(message.from_user.id)))
+    get_len = await get_button_labels(message.from_user.id)
 
-        await message.answer("Выберите категорию:", reply_markup=keyboard)
-        await state.set_state("waiting_for_category_name_for_deletion")
+    try:
+        if len(get_len) > 0:
+            keyboard = ReplyKeyboardMarkup(resize_keyboard=True, row_width=4, selective=True).add(
+                *(KeyboardButton(text=label) for label in await get_button_labels(message.from_user.id)))
+
+            await message.answer("Выберите категорию:", reply_markup=keyboard)
+            await state.set_state("waiting_for_category_name_for_deletion")
+        else:
+            await message.answer('Вам нечего удалять!')
     except Exception as e:
         logging.exception(f"Exception: {e} occurred while handling category deletion.")
+
+
+@dp.message_handler(lambda message: asyncio.ensure_future(is_in_buttons(message.from_user.id, message.text)))
+async def handle_category_click_wrapper(message: types.Message):
+    await handle_category_click(message)
 
 
 @dp.callback_query_handler(lambda c: c.data and c.data.startswith('delete_task:'))
@@ -182,17 +193,20 @@ async def process_callback_delete_task(callback_query: types.CallbackQuery):
 @dp.message_handler(state="waiting_for_task_description")
 async def create_task_handler(message: types.Message, state: FSMContext):
     task_description = message.text
-    json_data = {
-        'title': task_description,
-        'user_id': message.from_user.id
-    }
-    async with aiohttp.ClientSession() as session:
-        async with session.post('http://127.0.0.1:8000/task/create', json=json_data) as response:
-            r = await response.json()
-    await message.answer(f"Таск создан! {task_description}\n\nХотите добавить категорию?",
-                         reply_markup=add_category_keyboard)
-    await state.update_data(task_id=r['id'], user_id=r['user_id'])
-    await state.set_state(CreateTask.waiting_for_add_category)
+    if task_description not in ['Все задачи', 'Категории', 'Создать таск']:
+        json_data = {
+            'title': task_description,
+            'user_id': message.from_user.id
+        }
+        async with aiohttp.ClientSession() as session:
+            async with session.post('http://127.0.0.1:8000/task/create', json=json_data) as response:
+                r = await response.json()
+        await message.answer(f"Таск создан! {task_description}\n\nХотите добавить категорию?",
+                             reply_markup=add_category_keyboard)
+        await state.update_data(task_id=r['id'], user_id=r['user_id'])
+        await state.set_state(CreateTask.waiting_for_add_category)
+    else:
+        await message.answer('Недопустимая название таска!')
 
 
 @dp.callback_query_handler(lambda query: query.data == "add_category:yes", state=CreateTask.waiting_for_add_category)
@@ -207,6 +221,8 @@ async def add_category_handler(query: types.CallbackQuery, state: FSMContext):
     await state.set_state(CreateTask.waiting_for_category_selection)
 
 
+# todo проверка на то, что удалялись существубщие категории, проверка при создании на кнопки, чтоб они не создавались
+
 @dp.message_handler(state="waiting_for_category_name_for_deletion")
 async def handle_category_name_for_deletion(message: types.Message, state: FSMContext) -> None:
     """
@@ -218,12 +234,15 @@ async def handle_category_name_for_deletion(message: types.Message, state: FSMCo
     try:
         user_id = message.from_user.id
         category_name = message.text
+        if category_name in await get_button_labels(user_id):
 
-        await delete_category(user_id, category_name)  # Call the function to delete the category
-        await message.answer(f"Категория {category_name} удалена.")
-        keyboard = await send_category_keyboard(message.from_user.id)
-        await bot.send_message(chat_id=message.chat.id, text="Выберете категорию:", reply_markup=keyboard)
-        await state.finish()
+            await delete_category(user_id, category_name)  # Call the function to delete the category
+            await message.answer(f"Категория {category_name} удалена.")
+            keyboard = await send_category_keyboard(message.from_user.id)
+            await bot.send_message(chat_id=message.chat.id, text="Выберете категорию:", reply_markup=keyboard)
+            await state.finish()
+        else:
+            await message.answer('Такой категории не существует!')
     except Exception as e:
         logging.exception(f"Exception: {e} occurred while handling category deletion.")
 
@@ -247,6 +266,7 @@ async def process_category_name(message: types.Message, state: FSMContext):
 @dp.callback_query_handler(lambda query: query.data == "add_category:no", state=CreateTask.waiting_for_add_category)
 async def no_category_handler(query: types.CallbackQuery, state: FSMContext):
     await query.message.delete()
+    await state.finish()
     await bot.send_message(query.message.chat.id, "Главное меню", reply_markup=main_keyboard)
